[fix]: Moved a test to a move appropriate location
